Move en-savoir-plus/ to ir/
